using "long long" for timestamps, else we overrun on 32-bit systems.

+++ src/bare/apps/trafficshaper/traffic_queue.c
@@ -191,7 +191,7 @@ get_oldest_node(struct Traffic_Queue *queue,
 
         char str_time_in_us_in_Q[20] = {0};
 
-        unsigned long time_in_us_in_Q =
+        unsigned long long time_in_us_in_Q =
             pelion_get_current_timestamp_us() - result->time_us;
 
         pelion_convert_numeric_to_float(time_in_us_in_Q,
